backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
scikit-learn
nltk
pdfminer.six
python-dotenv
numpy